+++ tests/test_inline_references.py
@@ -9,8 +9,8 @@ from docutils.parsers.rst import states
 from docutils.utils import Reporter
 from sphinx.testing.util import SphinxTestApp
 
-from sphinxcontrib.typst.translator import TypstTranslator
-from sphinxcontrib.typst.writer import TypstWriter
+from typsphinx.translator import TypstTranslator
+from typsphinx.writer import TypstWriter
 
 
 def create_document():